+++ client/src/RecipeForm.css
@@ -1,0 +1,37 @@
+.recipe-form-overlay {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+}
+
+.recipe-form {
+    display: flex;
+    flex-direction: row;
+    align-items: top;
+    height: 80vh;
+    min-width: 60vw;
+    max-width: 80vw;
+    background-color: rgb(255, 255, 255);
+    padding: 40px;
+    border-radius: 8px;
+}
+
+.form-left-column {
+    overflow: overlay;
+    scrollbar-width: none;
+    display: inline-block;
+    vertical-align: top;
+    max-width: 20vw;
+    border-right: solid 1px black;
+}
+
+.form-right-column {
+    position: relative;
+    overflow: overlay;
+    scrollbar-width: none;
+    height: auto;
+    display: inline-block;
+    max-width: 50vw;
+    vertical-align: top;
+}
