+++ client/src/RecipeList.css
@@ -2,16 +2,23 @@
     display: flex;
     flex-direction: column;
     align-items: flex-start;
+    border-radius: 8px;
 }
 
 .recipe-item {
     background-color: rgb(241, 241, 241);
-    border: 1px solid #000000;
-    border-radius: 8px;
+    border-radius: 16px;
     margin: 20px;
     padding: 20px;
     max-width: 70vw;
     width: 50vw;
+    transition-duration: 0.1s;
+    /* border: 1px solid #000000; */
+    box-shadow: 0px 0px 5px rgba(100, 100, 100, 0.5);
+}
+
+.recipe-item:hover {
+    box-shadow: 0px 0px 15px rgba(0, 0, 0, 0.5);
 }
 
 h1 {
@@ -22,6 +29,7 @@ h1 {
 img {
     height: 170px;
     border-radius: 8px;
+    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.5);
 }
 
 button {
