+++ client/src/RecipePanel.css
@@ -37,11 +37,8 @@
 }
 .panel-right-column {
   position: relative;
-  /* Make the container for buttons relative */
   overflow: overlay;
-  /* Ensure overflow content is scrollable */
   scrollbar-width: none;
-  /* Hide the scrollbar */
   width: 40%;
   height: auto;
   display: inline-block;
@@ -80,6 +77,7 @@
   max-height: 40vh;
   object-fit:cover;
   border-radius: 8px;
+  
 }
 
 .panel-recipe-details {
@@ -135,3 +133,4 @@
   color: black;
   border: none;
 }
+
